pyproject.toml
[tool.poetry]
name = "nonebot-plugin-helldivers"
version = "1.1.3"
description = "为了超级地球！"
authors = ["SherkeyXD <253294679@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_helldivers"}]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
nonebot2 = "^2.1.0"
nonebot_plugin_htmlrender = "^0.3.0"
nonebot-plugin-send-anything-anywhere = "^0.6.0"

[tool.poetry.group.test.dependencies]
nonebug = "^0.3.7"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
nonebot-adapter-onebot = "^2.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"